move repo to local
